requirements.txt
Django==3.1.3
Faker==4.0.2
PyMySQL==0.10.1
djangorestframework==3.11.0
graphene-django==2.7.1
mysqlclient==1.4.6

flake8==3.8.4
isort==5.6.4
jedi==0.17.2

pydantic==1.6.1
